Move video frame rendering from GlVideoRenderer into GlFrameRenderFilter.
