+++ litr-demo/src/main/java/com/linkedin/android/litr/demo/data/TransformationPresenter.java
@@ -29,7 +29,7 @@ import com.linkedin.android.litr.codec.MediaCodecEncoder;
 import com.linkedin.android.litr.exception.MediaTransformationException;
 import com.linkedin.android.litr.filter.GlFilter;
 import com.linkedin.android.litr.filter.GlFrameRenderFilter;
-import com.linkedin.android.litr.filter.video.gl.FreeTransformFrameRenderFilter;
+import com.linkedin.android.litr.filter.video.gl.GlVideoFrameRenderFilter;
 import com.linkedin.android.litr.io.MediaExtractorMediaSource;
 import com.linkedin.android.litr.io.MediaMuxerMediaTarget;
 import com.linkedin.android.litr.io.MediaSource;
@@ -176,7 +176,7 @@ public class TransformationPresenter {
                         filters.add(backgroundImageFilter);
                     }
 
-                    GlFrameRenderFilter frameRenderFilter = new FreeTransformFrameRenderFilter(new PointF(0.25f, 0.25f), new PointF(0.65f, 0.55f), 30);
+                    GlFrameRenderFilter frameRenderFilter = new GlVideoFrameRenderFilter(new PointF(0.25f, 0.25f), new PointF(0.65f, 0.55f), 30);
                     filters.add(frameRenderFilter);
 
                     trackTransformBuilder.setRenderer(new GlVideoRenderer(filters));
