+++ litr/src/main/java/com/linkedin/android/litr/filter/util/GlFilterUtil.java
