+++ litr/src/main/java/com/linkedin/android/litr/render/GlVideoRenderer.java
@@ -26,17 +26,13 @@ import android.opengl.Matrix;
 import android.os.Build;
 import android.view.Surface;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.linkedin.android.litr.codec.Frame;
 import com.linkedin.android.litr.filter.GlFilter;
 import com.linkedin.android.litr.filter.GlFrameRenderFilter;
-import com.linkedin.android.litr.filter.video.gl.ScaleToFitGlFrameRenderFilter;
+import com.linkedin.android.litr.filter.video.gl.GlVideoFrameRenderFilter;
 
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-import java.nio.FloatBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -51,75 +47,24 @@ public class GlVideoRenderer implements Renderer {
                                                  ? MediaFormat.KEY_ROTATION
                                                  : "rotation-degrees";
 
-    private static final int FLOAT_SIZE_BYTES = 4;
-    private static final int TRIANGLE_VERTICES_DATA_STRIDE_BYTES = 5 * FLOAT_SIZE_BYTES;
-    private static final int TRIANGLE_VERTICES_DATA_POS_OFFSET = 0;
-    private static final int TRIANGLE_VERTICES_DATA_UV_OFFSET = 3;
-
-    // shaders
-    private static final String VERTEX_SHADER =
-            "uniform mat4 uMVPMatrix;\n" +
-                    "uniform mat4 uSTMatrix;\n" +
-                    "attribute vec4 aPosition;\n" +
-                    "attribute vec4 aTextureCoord;\n" +
-                    "varying vec2 vTextureCoord;\n" +
-                    "void main() {\n" +
-                    "  gl_Position = uMVPMatrix * aPosition;\n" +
-                    "  vTextureCoord = (uSTMatrix * aTextureCoord).xy;\n" +
-                    "}\n";
-    private static final String FRAGMENT_SHADER =
-            "#extension GL_OES_EGL_image_external : require\n" +
-                    "precision mediump float;\n" +      // highp here doesn't seem to matter
-                    "varying vec2 vTextureCoord;\n" +
-                    "uniform samplerExternalOES sTexture;\n" +
-                    "void main() {\n" +
-                    "  gl_FragColor = texture2D(sTexture, vTextureCoord);\n" +
-                    "}\n";
-
     private VideoRenderInputSurface inputSurface;
     private VideoRenderOutputSurface outputSurface;
     private List<GlFilter> filters;
-    private final String fragmentShader;
-
-    private FloatBuffer triangleVertices;
-    private final float[] triangleVerticesData = {
-            // X, Y, Z, U, V
-            -1.0f, -1.0f, 0, 0.f, 0.f,
-            1.0f, -1.0f, 0, 1.f, 0.f,
-            -1.0f, 1.0f, 0, 0.f, 1.f,
-            1.0f, 1.0f, 0, 1.f, 1.f,
-    };
 
     private float[] mvpMatrix = new float[16];
-
-    private int glProgram;
-    private int aPositionHandle;
-    private int aTextureHandle;
 
     private boolean inputSurfaceTextureInitialized;
 
     /**
      * Create an instance of GlVideoRenderer. If filter list has a {@link GlFrameRenderFilter}, that filter
-     * will be used to render video frames. Otherwise, default {@link ScaleToFitGlFrameRenderFilter}
+     * will be used to render video frames. Otherwise, default {@link GlVideoFrameRenderFilter}
      * will be used at lowest Z level to render video frames.
      * @param filters optional list of OpenGL filters to applied to output video frames
      */
     public GlVideoRenderer(@Nullable List<GlFilter> filters) {
-        this(FRAGMENT_SHADER, filters);
-    }
-
-    /**
-     * Create an instance of GlVideoRenderer. If filter list has a {@link GlFrameRenderFilter}, that filter
-     * will be used to render video frames. Otherwise, default {@link ScaleToFitGlFrameRenderFilter}
-     * will be used at lowest Z level to render video frames.
-     * @param fragmentShader allow to override default fragment shade to apply some customizations
-     * @param filters optional list of OpenGL filters to applied to output video frames
-     */
-    public GlVideoRenderer(@NonNull String fragmentShader, @Nullable List<GlFilter> filters) {
-        this.fragmentShader = fragmentShader;
         this.filters = new ArrayList<>();
         if (filters == null) {
-            this.filters.add(new ScaleToFitGlFrameRenderFilter());
+            this.filters.add(new GlVideoFrameRenderFilter());
             return;
         }
 
@@ -132,7 +77,7 @@ public class GlVideoRenderer implements Renderer {
         }
         if (!hasFrameRenderFilter) {
             // if client provided filters don't have a frame render filter, insert default frame filter
-            this.filters.add(new ScaleToFitGlFrameRenderFilter());
+            this.filters.add(new GlVideoFrameRenderFilter());
         }
         this.filters.addAll(filters);
     }
@@ -145,11 +90,6 @@ public class GlVideoRenderer implements Renderer {
         if (targetMediaFormat == null) {
             throw new IllegalArgumentException("GlVideoRenderer requires target media format");
         }
-
-        triangleVertices = ByteBuffer.allocateDirect(
-                triangleVerticesData.length * FLOAT_SIZE_BYTES)
-                .order(ByteOrder.nativeOrder()).asFloatBuffer();
-        triangleVertices.put(triangleVerticesData).position(0);
 
         // prioritize target video rotation value, fall back to source video rotation value
         int rotation = 0;
@@ -167,7 +107,6 @@ public class GlVideoRenderer implements Renderer {
 
         inputSurface = new VideoRenderInputSurface();
         initMvpMatrix(rotation, aspectRatio);
-        initGl();
 
         for (GlFilter filter : filters) {
             filter.init(Arrays.copyOf(mvpMatrix, mvpMatrix.length), 0);
@@ -210,19 +149,6 @@ public class GlVideoRenderer implements Renderer {
 
         GLES20.glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
         GLES20.glClear(GLES20.GL_DEPTH_BUFFER_BIT | GLES20.GL_COLOR_BUFFER_BIT);
-
-        triangleVertices.position(TRIANGLE_VERTICES_DATA_POS_OFFSET);
-        GLES20.glVertexAttribPointer(aPositionHandle, 3, GLES20.GL_FLOAT, false,
-                TRIANGLE_VERTICES_DATA_STRIDE_BYTES, triangleVertices);
-        GlRenderUtils.checkGlError("glVertexAttribPointer maPosition");
-        GLES20.glEnableVertexAttribArray(aPositionHandle);
-        GlRenderUtils.checkGlError("glEnableVertexAttribArray aPositionHandle");
-        triangleVertices.position(TRIANGLE_VERTICES_DATA_UV_OFFSET);
-        GLES20.glVertexAttribPointer(aTextureHandle, 2, GLES20.GL_FLOAT, false,
-                TRIANGLE_VERTICES_DATA_STRIDE_BYTES, triangleVertices);
-        GlRenderUtils.checkGlError("glVertexAttribPointer aTextureHandle");
-        GLES20.glEnableVertexAttribArray(aTextureHandle);
-        GlRenderUtils.checkGlError("glEnableVertexAttribArray aTextureHandle");
 
         for (GlFilter filter : filters) {
             filter.apply(presentationTimeNs);
@@ -271,26 +197,6 @@ public class GlVideoRenderer implements Renderer {
 
         Matrix.setIdentityM(mvpMatrix, 0);
         Matrix.multiplyMM(mvpMatrix, 0, projectionMatrix, 0, viewMatrix, 0);
-    }
-
-    /**
-     * Initializes GL state.  Call this after the EGL surface has been created and made current.
-     */
-    private void initGl() {
-        glProgram = GlRenderUtils.createProgram(VERTEX_SHADER, fragmentShader);
-        if (glProgram == 0) {
-            throw new RuntimeException("failed creating glProgram");
-        }
-        aPositionHandle = GLES20.glGetAttribLocation(glProgram, "aPosition");
-        GlRenderUtils.checkGlError("glGetAttribLocation aPosition");
-        if (aPositionHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for aPosition");
-        }
-        aTextureHandle = GLES20.glGetAttribLocation(glProgram, "aTextureCoord");
-        GlRenderUtils.checkGlError("glGetAttribLocation aTextureCoord");
-        if (aTextureHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for aTextureCoord");
-        }
     }
 
     private void initInputSurfaceTexture() {
