+++ litr/src/main/java/com/linkedin/android/litr/shader/FragmentShaders.java
@@ -2,7 +2,17 @@ package com.linkedin.android.litr.shader;
 
 public class FragmentShaders {
 
-    public static final String GRAYSCALE_SHADER = "#extension GL_OES_EGL_image_external : require\n" +
+    public static final String DEFAULT_SHADER =
+            "#extension GL_OES_EGL_image_external : require\n" +
+            "precision mediump float;\n" +      // highp here doesn't seem to matter
+            "varying vec2 vTextureCoord;\n" +
+            "uniform samplerExternalOES sTexture;\n" +
+            "void main() {\n" +
+            "  gl_FragColor = texture2D(sTexture, vTextureCoord);\n" +
+            "}\n";
+
+    public static final String GRAYSCALE_SHADER =
+            "#extension GL_OES_EGL_image_external : require\n" +
             "precision mediump float;\n" +
             "varying vec2 vTextureCoord;\n" +
             "uniform samplerExternalOES sTexture;\n" +
