+++ litr/src/main/java/com/linkedin/android/litr/shader/VertexShaders.java
@@ -1,0 +1,15 @@
+package com.linkedin.android.litr.shader;
+
+public class VertexShaders {
+
+    public static final String DEFAULT_SHADER =
+            "uniform mat4 uMVPMatrix;\n" +
+            "uniform mat4 uSTMatrix;\n" +
+            "attribute vec4 aPosition;\n" +
+            "attribute vec4 aTextureCoord;\n" +
+            "varying vec2 vTextureCoord;\n" +
+            "void main() {\n" +
+            "  gl_Position = uMVPMatrix * aPosition;\n" +
+            "  vTextureCoord = (uSTMatrix * aTextureCoord).xy;\n" +
+            "}\n";
+}
